Laba 3.1.7

+++ app/Http/Controllers/CommentController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Comment; // импорт
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class CommentController extends Controller
 {
@@ -17,7 +19,7 @@ class CommentController extends Controller
         $comment = new Comment;
         $comment->title = $request->title;
         $comment->text = $request->text;
-        $comment->author_id = 1;
+        $comment->author_id = Auth::id(); // Назначение автора
         $comment->article_id = $request->article_id;
         $comment->save();
     
@@ -26,6 +28,7 @@ class CommentController extends Controller
 
     public function delete($comment_id) {
         $comment = Comment::findOrFail($comment_id);
+        Gate::authorize('comment', $comment); // Проверка на право доступа
         $article_id = $comment->article_id;
         $comment->delete();
         return redirect()->route('article.show', ['article' => $article_id]);
@@ -38,6 +41,7 @@ class CommentController extends Controller
         ]);
     
         $comment = Comment::findOrFail($comment_id);
+        Gate::authorize('comment', $comment); // Проверка на право доступа
         $comment->title = $request->title;
         $comment->text = $request->text;
         $comment->save();
@@ -47,6 +51,7 @@ class CommentController extends Controller
 
     public function edit($comment_id) {
         $comment = Comment::findOrFail($comment_id);
+        Gate::authorize('comment', $comment); // Проверка на право 
         return view('comment.edit_comment', ['comment' => $comment]);
     }
